Resistor and Capacitor field data type change, ResistorBarcode added

+++ src/main/java/com/example/electric_storage/api/capacitor/CapacitorRequest.java
@@ -4,9 +4,9 @@ import com.example.electric_storage.domain.capacitor.CapacitorType;
 import com.example.electric_storage.infrastructure.common.ValueMultiplier;
 
 public record CapacitorRequest(
-    int capacitance,
+    double capacitance,
     ValueMultiplier multiplier,
-    float tolerance,
-    float power,
-    int impedance,
+    double tolerance,
+    double power,
+    double impedance,
     CapacitorType capacitorType) {}
