+++ src/main/java/com/example/electric_storage/api/resistor/ResistorRequest.java
@@ -3,4 +3,4 @@ package com.example.electric_storage.api.resistor;
 import com.example.electric_storage.infrastructure.common.ValueMultiplier;
 
 public record ResistorRequest(
-    int resistance, ValueMultiplier multiplier, float tolerance, float power) {}
+    double resistance, ValueMultiplier multiplier, double tolerance, double power) {}
